+++ packages/core/test/fakeLayout.test.ts
@@ -1,4 +1,4 @@
-import { FakeLayoutEngine } from "../src/index.js";
+import { FakeLayoutEngine } from "../src/testing.js";
 import { layoutContract } from "./layoutContract.js";
 
 layoutContract("FakeLayoutEngine", () => new FakeLayoutEngine());